build: Implement React-Router for client side routing
